+++ gazebo/gui/model/ModelEditorPalette.cc
@@ -178,15 +178,23 @@ ModelEditorPalette::ModelEditorPalette(QWidget *_parent)
   this->modelTreeWidget->setVerticalScrollMode(
       QAbstractItemView::ScrollPerPixel);
 
+  // Model Plugins
+  this->modelPluginsItem = new QTreeWidgetItem(static_cast<QTreeWidgetItem*>(0),
+      QStringList(QString("%1").arg(tr("Model Plugins"))));
+  this->modelPluginsItem->setData(0, Qt::UserRole,
+      QVariant(tr("Model Plugins")));
+  QFont headerFont = this->modelPluginsItem->font(0);
+  headerFont.setBold(true);
+  headerFont.setPointSize(1.0 * headerFont.pointSize());
+  this->modelPluginsItem->setFont(0, headerFont);
+  this->modelTreeWidget->addTopLevelItem(this->modelPluginsItem);
+
   // Links
   this->linksItem = new QTreeWidgetItem(
       static_cast<QTreeWidgetItem *>(0),
       QStringList(QString("%1").arg(tr("Links"))));
   this->linksItem->setData(0, Qt::UserRole, QVariant(tr("Links")));
-  QFont linksFont = this->linksItem->font(0);
-  linksFont.setBold(true);
-  linksFont.setPointSize(1.1 * linksFont.pointSize());
-  this->linksItem->setFont(0, linksFont);
+  this->linksItem->setFont(0, headerFont);
   this->modelTreeWidget->addTopLevelItem(this->linksItem);
 
   // Joints
@@ -194,7 +202,7 @@ ModelEditorPalette::ModelEditorPalette(QWidget *_parent)
       static_cast<QTreeWidgetItem*>(0),
       QStringList(QString("%1").arg(tr("Joints"))));
   this->jointsItem->setData(0, Qt::UserRole, QVariant(tr("Joints")));
-  this->jointsItem->setFont(0, linksFont);
+  this->jointsItem->setFont(0, headerFont);
   this->modelTreeWidget->addTopLevelItem(this->jointsItem);
 
   connect(this->modelTreeWidget,
@@ -266,6 +274,10 @@ ModelEditorPalette::ModelEditorPalette(QWidget *_parent)
   this->connections.push_back(
       gui::model::Events::ConnectJointInserted(
       boost::bind(&ModelEditorPalette::OnJointInserted, this, _1, _2, _3, _4)));
+
+  this->connections.push_back(
+      gui::model::Events::ConnectModelPluginInserted(
+      boost::bind(&ModelEditorPalette::OnModelPluginInserted, this, _1)));
 
   this->connections.push_back(
       gui::model::Events::ConnectLinkRemoved(
@@ -547,6 +559,8 @@ void ModelEditorPalette::OnItemDoubleClicked(QTreeWidgetItem *_item,
       gui::model::Events::openLinkInspector(name);
     else if (type == "Joint")
       gui::model::Events::openJointInspector(name);
+    else if (type == "Model Plugin")
+      gui::model::Events::openModelPluginInspector(name);
   }
 }
 
@@ -655,6 +669,21 @@ void ModelEditorPalette::OnJointInserted(const std::string &_jointId,
 }
 
 /////////////////////////////////////////////////
+void ModelEditorPalette::OnModelPluginInserted(
+    const std::string &_modelPluginName)
+{
+  QTreeWidgetItem *newModelPluginItem = new QTreeWidgetItem(
+      this->modelPluginsItem, QStringList(QString("%1").arg(
+      QString::fromStdString(_modelPluginName))));
+
+  newModelPluginItem->setData(0, Qt::UserRole, _modelPluginName.c_str());
+  newModelPluginItem->setData(1, Qt::UserRole, "Model Plugin");
+  this->modelTreeWidget->addTopLevelItem(newModelPluginItem);
+
+  this->modelPluginsItem->setExpanded(true);
+}
+
+/////////////////////////////////////////////////
 void ModelEditorPalette::OnLinkRemoved(const std::string &_linkId)
 {
   std::unique_lock<std::recursive_mutex> lock(this->updateMutex);
@@ -700,6 +729,8 @@ void ModelEditorPalette::ClearModelTree()
   this->linksItem->takeChildren();
   // Remove all joints
   this->jointsItem->takeChildren();
+  // Remove all model plugins
+  this->modelPluginsItem->takeChildren();
 }
 
 /////////////////////////////////////////////////